[overal] Turned on all warnings and fixed all warnings (nameless structs warning is suppressed). This also includes renaming variables and predeclaring classes [deps] Added stb/zlib and rapidjson as proper dependencies (not included in project anymore) [ogc] initSceneSurface now takes Vec2u res. Added templated Graphics::destroy function (as a helper to set the GraphicsObject* to nullptr). Moved spvhelper [ogc vk] Log prints have now been restructured and decreased

+++ ogc/src/graphics/format/oisb.cpp
@@ -189,9 +189,9 @@ SBFile oiSB::convert(ShaderBufferInfo &info, SLFile *names) {
 		}
 
 		if (elem.format != TextureFormat::Undefined)
-			file.vars.push_back(SBVar(names->lookup(elem.name), info.lookup(elem.parent), elem.offset, arrayIndex, (u8) elem.format.getValue(), (u8) elem.flags));
+			file.vars.push_back(SBVar((u16) names->lookup(elem.name), (u16) info.lookup(elem.parent), elem.offset, arrayIndex, (u8) elem.format.getValue(), (u8) elem.flags));
 		else 
-			file.structs.push_back(SBStruct(names->lookup(elem.name), info.lookup(elem.parent), elem.offset, arrayIndex, (u8) elem.flags, elem.length));
+			file.structs.push_back(SBStruct((u16) names->lookup(elem.name), (u16) info.lookup(elem.parent), elem.offset, arrayIndex, (u8) elem.flags, elem.length));
 
 	}
